Add funtionality to removeTask function in add-remove.js
