+++ src/display.js
@@ -1,5 +1,5 @@
 // display.js
-import { Task, tasks } from './index.js';
+import { tasks } from './add-remove.js';
 import arrow from './assets/arrow.svg';
 
 const wrapper = document.getElementById('wrapper');
@@ -54,10 +54,12 @@ export function renderTodoList() {
   const todoListContainer = document.getElementById('todo');
   todoListContainer.innerHTML = '';
   tasks.sort((a, b) => a.index - b.index);
-  tasks.forEach((task) => {
+  
+  tasks.forEach((task, index) => { // Update id when rendering the todo list
+    task.index = index; // Update the index of the task
     const listItem = document.createElement('div');
     listItem.className = 'todo-item';
-    listItem.id = task.index;
+    listItem.id = task.index; // Set the id as the index
     listItem.innerHTML = `
       <div class="checkbox"><i class="fa-regular fa-square"></i></div>
       <div class="description">${task.description}</div>
