+++ src/index.js
@@ -3,18 +3,12 @@ import arrow from './assets/arrow.svg';
 import { addTask, removeTask } from './add-remove.js';
 import { renderTodoList } from './display.js';
 
-// create Class of task 
-export class Task {
-  constructor(completed, description, index) {
-    this.completed = completed;
-    this.description = description;
-    this.index = index;
-  }
-}
+document.addEventListener('DOMContentLoaded', () => {
+  addTask();
+  renderTodoList();
+  removeTask();
+});
 
-export const tasks = localStorage.getItem('tasks') ? JSON.parse(localStorage.getItem('tasks')) : [];
 
-addTask();
-renderTodoList();
-removeTask();
+
