+++ src/template.html
@@ -13,4 +13,5 @@
       <div id="wrapper"></div>
     </section>
   </body>
+  <script defer src="bundle.js"></script>
 </html>
